California csv file AND Jupiter Notebook
In the Jupiter notebook, I clean those 2 nasty columns.  I'm sure that there was a "cleaner" code that I could have used -- however, there was a significant lack of unioformity in the data of those 2 columns, and hence the uglier code.  At the very bottom, I have a "Address_misc" column that contains the remainder data of the original "Premise Address" column.  

So all we need to do is choose how many columns to include for the final final data frame and that's where I left off.

Next we need to pull in census data and make sure that "Zip Code" is a part of that csv file.  

We also need to create a MySQL notebook as well as then continuing with the Jupiter Notebook to finish the joining code.

See you in the morning!
-Kevin